code formating for csv export
