+++ frontend/src/components/userDetail/elementsMapped.js
@@ -12,6 +12,7 @@ import { Button } from '../button';
 import { Line } from 'react-chartjs-2';
 import Select from 'react-select';
 import moment from 'moment';
+import { exporttoCSVFile } from '../../network/genericCSVExport';
 import {
   ClockIcon,
   RoadIcon,
@@ -259,31 +260,6 @@ export const TaskStats = ({ userStats, username }) => {
     },
   ];
 
-  function convertArrayOfObjectsToCSV(array) {
-    let result;
-
-    const columnDelimiter = ',';
-    const lineDelimiter = '\n';
-    const keys = Object.keys(array[0]);
-
-    result = '';
-    result += keys.join(columnDelimiter);
-    result += lineDelimiter;
-
-    array.forEach((item) => {
-      let ctr = 0;
-      keys.forEach((key) => {
-        if (ctr > 0) result += columnDelimiter;
-
-        result += item[key];
-
-        ctr++;
-      });
-      result += lineDelimiter;
-    });
-
-    return result;
-  }
   const state = {
     labels: graphDays,
     datasets: [
@@ -305,22 +281,6 @@ export const TaskStats = ({ userStats, username }) => {
         data: graphValidationValues,
       },
     ],
-  };
-
-  const exporttoCsv = (dataArray, fileName) => {
-    const link = document.createElement('a');
-    let csv = convertArrayOfObjectsToCSV(dataArray);
-    if (csv == null) return;
-
-    const filename = userName + '-' + fileName + 'export.csv';
-
-    if (!csv.match(/^data:text\/csv/i)) {
-      csv = `data:text/csv;charset=utf-8,${csv}`;
-    }
-
-    link.setAttribute('href', encodeURI(csv));
-    link.setAttribute('download', filename);
-    link.click();
   };
 
   function submitSelected(Values) {
@@ -493,7 +453,7 @@ export const TaskStats = ({ userStats, username }) => {
                 <p>
                   <Button
                     className="bg-red white"
-                    onClick={() => exporttoCsv(dataSummaryResponse, 'Summary')}
+                    onClick={() => exporttoCSVFile(dataSummaryResponse, 'Summary')}
                   >
                     Export Results
                   </Button>
@@ -531,7 +491,7 @@ export const TaskStats = ({ userStats, username }) => {
                 <p>
                   <Button
                     className="bg-red white"
-                    onClick={() => exporttoCsv(dataMappedResponse, 'Mapped')}
+                    onClick={() => exporttoCSVFile(dataMappedResponse, 'Mapped')}
                   >
                     Export Results
                   </Button>
@@ -551,7 +511,7 @@ export const TaskStats = ({ userStats, username }) => {
                 <p>
                   <Button
                     className="bg-red white"
-                    onClick={() => exporttoCsv(dataValidatedResponse, 'Validation')}
+                    onClick={() => exporttoCSVFile(dataValidatedResponse, 'Validation')}
                   >
                     Export Results
                   </Button>
